requirements.txt
fastapi
uvicorn
tensorflow
librosa
numpy
scikit-learn
matplotlib
seaborn
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
soundfile